git commit -m "refactor: use fixed-point representation for temperature in logger
- change temperature from float to int (scaled by 100)
- avoid float printf overhead in embedded environment
- fix incorrect temperature output formatting
"

+++ include/sensor_data.h
@@ -5,7 +5,7 @@
 
 typedef struct {
     uint32_t timestamp;
-    float temperature;
+    int temperature;
 } sensor_data_t;
 
 #endif
